M46: fix Round 3/4 audit findings + remaining Part-3 leads
Lands the actionable Round 3 & 4 findings and the unverified Part-3 leads
from docs/vfs-process-audit.md. Both arches build clean; x86 smoke is green
single-CPU and -smp 4 (only the two pre-existing failures remain:
nice-biasing flake, 32-bit userspace-on-ap).

Confirmed criticals:
- R3-1: vfs_open_flags double-decref UAF on the EMFILE path (NULL node so
  out: does not re-put the reference vfs_handle_release already dropped).
- R3-2: IPv6 ::1 loopback deferred via net_loopback_enqueue (TCP re-entrancy);
  ipv6_loopback_smoke pumps net_loopback_drain so the deferred cascade
  completes before sampling counters.
- R3-3: mqueue handle is now a validated table index, never a raw kernel
  pointer; global table spinlock added.
- R3-4/R3-5: ext4 uninitialized-extent high bit masked in lookup+merge;
  ext4_set_block no longer routes a block-mapped inode into ext4_add_extent.
- R4-1: RTC UIP spin bounded (both arches).
- R4-2/R4-3: multiboot2 mmap entry_size / tag size guards.
- R4-4: ELF64/32 e_phentsize + subtraction-form e_phoff bounds.

Leads / hardening:
- R3-10 ext2 dir-grow set-block helper + hole-read clamp; R3-11 aio ctx
  refcount (worker-vs-cleanup UAF); R3-12 CAS wake in kill/_process_group;
  R3-13 blk_flush_buffer keeps DIRTY on failed write-back; R3-14 32-bit fork
  stops COWing VMM_SHARED; R3-15 create/mkdir/symlink refcount underflow.
- R4-5 virtio-net drain cap; R4-6 ELF PT_INTERP/PT_DYNAMIC wrap guards;
  R4-7 PMM/module overflow guards; R4-9 NVMe count==0 + virtio_blk overflow;
  R4-10 ELF p_memsz ceiling; R4-11 PS/2 mouse packet_index bound.
- mqueue locking, futex exit-cleanup (scheduler_futex_cleanup_task), aio leads.
- M46-3: signal-death reparents children + orphan-pgrp signalling via shared
  reparent_children_and_signal_orphans() helper (no scheduler re-entry).

Follow-up batch:
- shm: owner/creator/root cred check on IPC_RMID/IPC_SET; SHM_RDONLY honored.
- R3-9: ext2/ext4 dir-walk rec_len bounds (lookup/readdir/remove/rename/add).
- R3-9: fat32 mount SPC/fat_count guards + self-referential FAT chain guard;
  dns answer-parse bounds.
- R4-8: bounded ELF strtab/symbol-name accessors (elf64_str_at_equals /
  elf64_copy_str) for the PIE dynamic-link resolver + DT_NEEDED reader.

Deferred (documented in the audit doc, need test infra): R3-6 vma_list lock,
R3-8 shm fork-nattch/detach-leak, unix lost-wakeups, journal crash-atomicity,
ext3/ext1/exfat/ntfs crafted-image hardening, futex feature-gaps.

+++ kernel/dev/virtio_blk.c
@@ -114,7 +114,7 @@ static int do_virtio_blk_req(struct virtio_blk_instance *inst, u64 lba,
   if (vblk_add_region(&inst->vq, &next_desc, &prev, (u64)(usize)&dma->req,
                       sizeof(struct virtio_blk_req), 0) < 0 ||
       vblk_add_region(&inst->vq, &next_desc, &prev, (u64)(usize)buffer,
-                      count * 512,
+                      (u64)count * 512,
                       type == VIRTIO_BLK_T_IN ? VRING_DESC_F_WRITE : 0) < 0 ||
       vblk_add_region(&inst->vq, &next_desc, &prev, (u64)(usize)&dma->status, 1,
                       VRING_DESC_F_WRITE) < 0) {
